refactor(domain): add repository interface
References:
 - https://proandroiddev.com/why-you-need-use-cases-interactors-142e8a6fe576
