+++ app/src/main/java/com/example/alpacamusclemaintenance/db/repository/PushUpRepositoryImpl.kt
@@ -2,11 +2,11 @@ package com.example.alpacamusclemaintenance.db.repository
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.map
-import com.example.alpacamusclemaintenance.data.pushup.PushUpRepository
 import com.example.alpacamusclemaintenance.db.dao.PushUpDao
 import com.example.alpacamusclemaintenance.db.dao.toPushUpEntity
 import com.example.alpacamusclemaintenance.db.dao.toPushUps
 import com.example.alpacamusclemaintenance.domain.pushup.PushUp
+import com.example.alpacamusclemaintenance.domain.pushup.PushUpRepository
 import javax.inject.Inject
 
 class PushUpRepositoryImpl @Inject constructor(
