+++ app/src/main/java/com/example/alpacamusclemaintenance/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.alpacamusclemaintenance.di
 
-import com.example.alpacamusclemaintenance.data.pushup.PushUpRepository
 import com.example.alpacamusclemaintenance.db.repository.PushUpRepositoryImpl
+import com.example.alpacamusclemaintenance.domain.pushup.PushUpRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
