+++ domain/build.gradle
@@ -43,6 +43,7 @@ apply from: '../dependencies-all.gradle'
 
 dependencies {
     core()
+    paging()
     testing()
 }
