+++ domain/src/main/java/com/example/alpacamusclemaintenance/domain/feed/FeedRepository.kt
@@ -1,0 +1,9 @@
+package com.example.alpacamusclemaintenance.domain.feed
+
+import androidx.paging.PagingData
+import kotlinx.coroutines.flow.Flow
+
+interface FeedRepository {
+
+    fun getFeedResultStream(query: String): Flow<PagingData<Feed>>
+}
