+++ domain/src/main/java/com/example/alpacamusclemaintenance/domain/home/HomeRepository.kt
@@ -1,0 +1,8 @@
+package com.example.alpacamusclemaintenance.domain.home
+
+import androidx.lifecycle.LiveData
+
+interface HomeRepository {
+
+    fun getHome(): LiveData<Home>
+}
